Move common visualizer keyframes into visualizer_keyframes.h
